separate the generic server functionality from the TCP specific parts
