+++ rodbus/src/server/task.rs
@@ -1,12 +1,6 @@
-use std::collections::BTreeMap;
-use std::sync::Arc;
-
 use futures::future::FutureExt;
 use futures::select;
-use log::{info, warn};
 use tokio::io::{AsyncRead, AsyncWrite, AsyncWriteExt};
-use tokio::net::TcpListener;
-use tokio::sync::Mutex;
 
 use crate::error::details::ExceptionCode;
 use crate::error::*;
@@ -20,93 +14,7 @@ use crate::types::{AddressRange, Indexed};
 use crate::util::cursor::ReadCursor;
 use crate::util::frame::{Frame, FrameFormatter, FrameHeader, FramedReader};
 
-struct SessionTracker {
-    max: usize,
-    id: u64,
-    sessions: BTreeMap<u64, tokio::sync::mpsc::Sender<()>>,
-}
-
-type SessionTrackerWrapper = Arc<Mutex<Box<SessionTracker>>>;
-
-impl SessionTracker {
-    fn new(max: usize) -> SessionTracker {
-        Self {
-            max,
-            id: 0,
-            sessions: BTreeMap::new(),
-        }
-    }
-
-    fn get_next_id(&mut self) -> u64 {
-        let ret = self.id;
-        self.id += 1;
-        ret
-    }
-
-    pub fn wrapped(max: usize) -> SessionTrackerWrapper {
-        Arc::new(Mutex::new(Box::new(Self::new(max))))
-    }
-
-    pub fn add(&mut self, sender: tokio::sync::mpsc::Sender<()>) -> u64 {
-        // TODO - this is so ugly. there's a nightly API on BTreeMap that has a remove_first
-        if !self.sessions.is_empty() && self.sessions.len() >= self.max {
-            let id = *self.sessions.keys().next().unwrap();
-            warn!("exceeded max connections, closing oldest session: {}", id);
-            // when the record drops, and there are no more senders,
-            // the other end will stop the task
-            self.sessions.remove(&id).unwrap();
-        }
-
-        let id = self.get_next_id();
-        self.sessions.insert(id, sender);
-        id
-    }
-
-    pub fn remove(&mut self, id: u64) {
-        self.sessions.remove(&id);
-    }
-}
-
-pub struct ServerTask<T: ServerHandler> {
-    listener: TcpListener,
-    handlers: ServerHandlerMap<T>,
-    tracker: SessionTrackerWrapper,
-}
-
-impl<T> ServerTask<T>
-where
-    T: ServerHandler,
-{
-    pub fn new(max_sessions: usize, listener: TcpListener, handlers: ServerHandlerMap<T>) -> Self {
-        Self {
-            listener,
-            handlers,
-            tracker: SessionTracker::wrapped(max_sessions),
-        }
-    }
-
-    pub async fn run(&mut self) -> std::io::Result<()> {
-        loop {
-            let (socket, addr) = self.listener.accept().await?;
-
-            let handlers = self.handlers.clone();
-            let tracker = self.tracker.clone();
-            let (tx, rx) = tokio::sync::mpsc::channel(1);
-
-            let id = self.tracker.lock().await.add(tx);
-
-            info!("accepted connection {} from: {}", id, addr);
-
-            tokio::spawn(async move {
-                SessionTask::new(socket, handlers, rx).run().await.ok();
-                info!("shutdown session: {}", id);
-                tracker.lock().await.remove(id);
-            });
-        }
-    }
-}
-
-struct SessionTask<T, U>
+pub(crate) struct SessionTask<T, U>
 where
     T: ServerHandler,
     U: AsyncRead + AsyncWrite + Unpin,
@@ -123,7 +31,7 @@ where
     T: ServerHandler,
     U: AsyncRead + AsyncWrite + Unpin,
 {
-    pub fn new(
+    pub(crate) fn new(
         io: U,
         handlers: ServerHandlerMap<T>,
         shutdown: tokio::sync::mpsc::Receiver<()>,
@@ -148,7 +56,7 @@ where
         Ok(())
     }
 
-    async fn run(&mut self) -> std::result::Result<(), Error> {
+    pub(crate) async fn run(&mut self) -> std::result::Result<(), Error> {
         loop {
             self.run_one().await?;
         }
@@ -173,7 +81,7 @@ where
         // if no addresses match, then don't respond
         let handler = match self.handlers.get(frame.header.unit_id) {
             None => {
-                warn!(
+                log::warn!(
                     "received frame for unmapped unit id: {}",
                     frame.header.unit_id.value
                 );
@@ -184,13 +92,13 @@ where
 
         let function = match cursor.read_u8() {
             Err(_) => {
-                warn!("received an empty frame");
+                log::warn!("received an empty frame");
                 return Ok(());
             }
             Ok(value) => match FunctionCode::get(value) {
                 Some(x) => x,
                 None => {
-                    warn!("received unknown function code: {}", value);
+                    log::warn!("received unknown function code: {}", value);
                     return self
                         .reply_with_exception(
                             frame.header,
@@ -209,7 +117,7 @@ where
             match function {
                 FunctionCode::ReadCoils => match AddressRange::parse(&mut cursor) {
                     Err(e) => {
-                        warn!("error parsing {:?} request: {}", function, e);
+                        log::warn!("error parsing {:?} request: {}", function, e);
                         self.writer.format(
                             frame.header,
                             &ADU::new(function.as_error(), &ExceptionCode::IllegalDataValue),
@@ -226,7 +134,7 @@ where
                 },
                 FunctionCode::ReadDiscreteInputs => match AddressRange::parse(&mut cursor) {
                     Err(e) => {
-                        warn!("error parsing {:?} request: {}", function, e);
+                        log::warn!("error parsing {:?} request: {}", function, e);
                         self.writer.format(
                             frame.header,
                             &ADU::new(function.as_error(), &ExceptionCode::IllegalDataValue),
@@ -243,7 +151,7 @@ where
                 },
                 FunctionCode::ReadHoldingRegisters => match AddressRange::parse(&mut cursor) {
                     Err(e) => {
-                        warn!("error parsing {:?} request: {}", function, e);
+                        log::warn!("error parsing {:?} request: {}", function, e);
                         self.writer.format(
                             frame.header,
                             &ADU::new(function.as_error(), &ExceptionCode::IllegalDataValue),
@@ -260,7 +168,7 @@ where
                 },
                 FunctionCode::ReadInputRegisters => match AddressRange::parse(&mut cursor) {
                     Err(e) => {
-                        warn!("error parsing {:?} request: {}", function, e);
+                        log::warn!("error parsing {:?} request: {}", function, e);
                         self.writer.format(
                             frame.header,
                             &ADU::new(function.as_error(), &ExceptionCode::IllegalDataValue),
@@ -277,7 +185,7 @@ where
                 },
                 FunctionCode::WriteSingleCoil => match Indexed::<bool>::parse(&mut cursor) {
                     Err(e) => {
-                        warn!("error parsing {:?} request: {}", function, e);
+                        log::warn!("error parsing {:?} request: {}", function, e);
                         self.writer.format(
                             frame.header,
                             &ADU::new(function.as_error(), &ExceptionCode::IllegalDataValue),
@@ -294,7 +202,7 @@ where
                 },
                 FunctionCode::WriteSingleRegister => match Indexed::<u16>::parse(&mut cursor) {
                     Err(e) => {
-                        warn!("error parsing {:?} request: {}", function, e);
+                        log::warn!("error parsing {:?} request: {}", function, e);
                         self.writer.format(
                             frame.header,
                             &ADU::new(function.as_error(), &ExceptionCode::IllegalDataValue),
@@ -311,7 +219,7 @@ where
                 },
                 FunctionCode::WriteMultipleCoils => match parse_write_multiple_coils(&mut cursor) {
                     Err(e) => {
-                        warn!("error parsing {:?} request: {}", function, e);
+                        log::warn!("error parsing {:?} request: {}", function, e);
                         self.writer.format(
                             frame.header,
                             &ADU::new(function.as_error(), &ExceptionCode::IllegalDataValue),
@@ -329,7 +237,7 @@ where
                 FunctionCode::WriteMultipleRegisters => {
                     match parse_write_multiple_registers(&mut cursor) {
                         Err(e) => {
-                            warn!("error parsing {:?} request: {}", function, e);
+                            log::warn!("error parsing {:?} request: {}", function, e);
                             self.writer.format(
                                 frame.header,
                                 &ADU::new(function.as_error(), &ExceptionCode::IllegalDataValue),
