+++ rodbus/src/tcp/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod client;
 pub(crate) mod frame;
+pub(crate) mod server;
